refactor(adapters): refreshCodexToken 을 dedicated 파일로 분리
issue #105 의 일부 — provider-adapters 의 refresh-token 위치 비대칭 해소.

기존:
- claude/refresh-claude-token.js 단독 파일
- codex 의 refreshCodexToken 은 exchange-codex-authorization-code.js 안에 inline
- 같은 일을 하는데 위치만 다름

변경:
- 신규 codex/refresh-codex-token.js 생성, refreshCodexToken 이동
- exchange-codex-authorization-code.js 는 authorization_code 교환만 보유
- codex/index.js 에 './refresh-codex-token.js' export 추가
- provider-adapters/README.md 의 codex 파일 트리 갱신

테스트:
- 신규 test/codex/refresh-codex-token.test.js (4건, 기존 exchange test
  로부터 verbatim 이동 — POST 형식 / no-rotation / rotation / 401 에러)
- exchange-codex-authorization-code.test.js 의 refreshCodexToken describe
  블록 제거 + import 정리

영향:
- public API 변화 없음 — 모든 소비자가 codex/index.js 를 통해 import
  (agent/services/codex-provider.js, agent/cli/doctor-helpers.js)
- 두 provider 가 'refresh-*-token.js' 동일한 파일 구조로 정렬

검증:
- 기존 758 + 본 분리 후도 동일 통과 (refresh 4건은 위치만 이동)
- npm test / npm run lint / npm run format:check / npm run build:types 모두 clean

본 PR 은 issue #105 (login-runner refactor) 의 provider-adapters 측면
일부. login-runner 본체 (saveLiveExchangeAccount 분해) 는 별도 작업.

patch changeset (사용자 가시 변화 0).
