+++ packages/provider-adapters/src/codex/exchange-codex-authorization-code.js
@@ -1,11 +1,12 @@
 /**
- * Codex (OpenAI) OAuth token 교환 (authorization_code / refresh_token).
+ * Codex (OpenAI) OAuth authorization_code 교환.
  *
- * 공통 transport 로직은 `../shared/oauth-token-endpoint.js`에 있다.
+ * refresh_token 흐름은 `./refresh-codex-token.js` (v0.3.0 분리, Claude 의
+ * refresh-claude-token.js 와 대칭). 공통 transport 로직은
+ * `../shared/oauth-token-endpoint.js`.
  *
  * 미해결:
  *   - client_secret 필요 여부 (현재는 public client 가정)
- *   - refresh token rotation 정책
  */
 
 import { CODEX_AUTH } from './codex-auth-constants.js';
@@ -56,40 +57,6 @@ export async function exchangeCodexAuthorizationCode({
       ...(clientSecret ? { client_secret: clientSecret } : {}),
     },
     errorPrefix: 'Token exchange failed',
-    fetchImpl,
-  });
-}
-
-/**
- * Refresh an access token using a refresh token.
- *
- * @param {{
- *   refreshToken: string,
- *   clientId?: string,
- *   clientSecret?: string,
- *   tokenEndpoint?: string,
- *   fetchImpl?: typeof fetch,
- * }} params
- * @returns {Promise<TokenResponse>}
- */
-export async function refreshCodexToken({
-  refreshToken,
-  clientId = CODEX_AUTH.observedClientId,
-  clientSecret,
-  tokenEndpoint = CODEX_AUTH.tokenEndpoint,
-  fetchImpl,
-}) {
-  return postToTokenEndpoint({
-    endpoint: tokenEndpoint,
-    encoding: 'form',
-    body: {
-      grant_type: 'refresh_token',
-      refresh_token: refreshToken,
-      client_id: clientId,
-      ...(clientSecret ? { client_secret: clientSecret } : {}),
-    },
-    errorPrefix: 'Token refresh failed',
-    fallbackRefreshToken: refreshToken,
     fetchImpl,
   });
 }
