+++ packages/provider-adapters/src/codex/index.js
@@ -3,3 +3,4 @@ export * from './fetch-codex-usage.js';
 export * from './codex-auth-constants.js';
 export * from './build-codex-authorization-url.js';
 export * from './exchange-codex-authorization-code.js';
+export * from './refresh-codex-token.js';
